+++ Jellyfin.Plugin.AccountRequest/ServiceRegistrator.cs
@@ -2,6 +2,7 @@ using Jellyfin.Plugin.AccountRequest.Api;
 using Jellyfin.Plugin.AccountRequest.Services;
 using MediaBrowser.Controller;
 using MediaBrowser.Controller.Plugins;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Jellyfin.Plugin.AccountRequest;
@@ -14,6 +15,7 @@ public class ServiceRegistrator : IPluginServiceRegistrator
     /// <inheritdoc />
     public void RegisterServices(IServiceCollection serviceCollection, IServerApplicationHost applicationHost)
     {
+        serviceCollection.AddSingleton<IStartupFilter, LoginInjectStartupFilter>();
         serviceCollection.AddSingleton<RequestStore>();
         serviceCollection.AddTransient<AccountRequestController>();
     }